Implemented createPhotoPrint ajax callback Added bootstrap to the project

+++ libppp/include/libppp.h
@@ -34,7 +34,7 @@ public:
     *  param[in] bufferData Pointer to the image data
     *  returns Image Id that can be used to recognise the image
     !*/
-    std::string setImage(const char *bufferData, int bufferLength);
+    std::string setImage(const char *bufferData, size_t bufferLength);
     
     std::string detectLandmarks(const std::string &imageId);
